refactor(watch): split watch.go into focused modules
Break up the ~680 line watch.go into four single-responsibility files:
- types.go: Event, FileState, DepContext, Graph, State structs
- daemon.go: Daemon struct and lifecycle (NewDaemon, Start, Stop)
- events.go: event loop, handling, logging, and delta calculations
- state.go: PID management, state persistence, daemon control

No functional changes - all tests pass.
